requirements.txt
torch>=2.1
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.2
PyYAML>=6.0.1
numpy>=1.26
tqdm>=4.66
